Refactor location of `Navbar`.
