+++ frontend/src/components/logo-mines.png
